requirements.txt
pandas
requests
streamlit
agno
python-dotenv
groq
duckduckgo-search
newspaper4k
lxml_html_clean
tabulate
Flask
flask-cors
flasgger
gunicorn
redis
